# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 羊咩快报 API
  version: 1.0.0
  description: 后端接口文档
paths:
...
  /content/{id}/tags:
    get:
      operationId: content_tags_retrieve
      description: 内容标签关联API，支持打标签和获取标签
      summary: 内容标签关联（打标签/获取标签）
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - content
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentTagRelationResponse'
          description: ''
    post:
      operationId: content_tags_create_2
      description: 内容标签关联API，支持打标签和获取标签
      summary: 内容标签关联（打标签/获取标签）
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - content
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentTagRelationResponse'
          description: ''
...
  /content/{id}/category/:
    get:
      operationId: content_category_list
      description: 获取指定分类下的内容，支持分页、排序、时长筛选。
      summary: 根据分类获取内容
      parameters:
      - in: query
        name: duration_max
        schema:
          type: integer
        description: 最大时长(秒)
      - in: query
        name: duration_min
        schema:
          type: integer
        description: 最小时长(秒)
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this 原始新闻.
        required: true
      - in: query
        name: page
        schema:
          type: integer
        description: 页码
      - in: query
        name: page_size
        schema:
          type: integer
        description: 每页数量
      - in: query
        name: sort
        schema:
          type: string
        description: 排序方式(latest|popular|hot|duration)
      tags:
      - content
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RawNews'
              examples:
                科技资讯内容示例:
                  value:
                  - id: 101
                    title: AI驱动的新闻推荐系统上线
                    content: 羊咩快报平台正式发布AI驱动的新闻推荐系统，提升用户个性化体验。
                    summary: AI推荐系统助力个性化资讯推送。
                    author: 新闻编辑部
                    source_url: https://news.yangmie.com/ai/101
                    image_urls:
                    - https://cdn.yangmie.com/images/ai-news-101.jpg
                    published_at: '2025-07-20T10:00:00Z'
                    category: 2
                    tags:
                    - AI
                    - 推荐
                    - 新闻
                    keywords:
                    - 人工智能
                    - 个性化推荐
                    - 新闻平台
                    sentiment_score: 0.92
                    importance_score: 0.95
                    is_processed: true
                    is_duplicate: false
                    language: zh-CN
                    created_at: '2025-07-20T10:00:00Z'
                    updated_at: '2025-07-20T10:00:00Z'
                  description: 真实业务场景下的科技资讯内容示例。
          description: ''
...
  /content/tags:
    get:
      operationId: content_tags_list
      description: 内容标签列表与创建API
      tags:
      - content
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentTag'
          description: ''
    post:
      operationId: content_tags_create
      description: 内容标签列表与创建API
      tags:
      - content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentTag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ContentTag'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ContentTag'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentTag'
          description: ''
...
components:
  schemas:
...
    ContentTag:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          title: 标签名称
          maxLength: 50
        slug:
          type: string
          title: URL标识
          maxLength: 50
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
          nullable: true
          title: 标签描述
        tag_type:
          allOf:
          - $ref: '#/components/schemas/TagTypeEnum'
          title: 标签类型
        color_code:
          type: string
          nullable: true
          title: 标签颜色
          maxLength: 7
        background_color:
          type: string
          nullable: true
          title: 背景色
          maxLength: 7
        icon:
          type: string
          nullable: true
          title: 图标
          maxLength: 50
        supported_content_types:
          title: 支持的内容类型
        usage_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          title: 使用次数
        content_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          title: 关联内容数
        popularity_score:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          title: 热度分值
        trending_score:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          title: 趋势分值
        is_active:
          type: boolean
          title: 是否启用
        is_featured:
          type: boolean
          title: 是否推荐
        is_trending:
          type: boolean
          title: 是否热门
        is_auto_generated:
          type: boolean
          title: 是否自动生成
        synonyms:
          nullable: true
          title: 同义词
        search_weight:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          title: 搜索权重
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: 创建时间
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: 更新时间
        category:
          type: integer
          nullable: true
          title: 所属分类
        related_tags:
          type: array
          items:
            type: integer
            title: 相关标签
          title: 相关标签
      required:
      - created_at
      - id
      - name
      - updated_at
    ContentTagRelationResponse:
      type: object
      properties:
        id:
          type: integer
        content_type:
          type: string
        object_id:
          type: integer
        tag:
          type: string
        relevance_score:
          type: number
          format: double
        is_auto_tagged:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
      - content_type
      - created_at
      - id
      - is_auto_tagged
      - object_id
      - relevance_score
      - tag
...
    RawNews:
      type: object
      description: 原始新闻内容序列化器
      properties:
        id:
          type: integer
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/TypeEnum'
          title: 内容类型
        source:
          type: string
          nullable: true
          title: 来源
          maxLength: 100
        title:
          type: string
          title: 标题
          maxLength: 500
        content:
          type: string
          title: 内容
        summary:
          type: string
          nullable: true
          title: 摘要
        author:
          type: string
          nullable: true
          title: 作者
          maxLength: 100
        source_url:
          type: string
          format: uri
          title: 原始链接
          maxLength: 500
        image_urls:
          nullable: true
          title: 图片链接
        published_at:
          type: string
          format: date-time
          title: 发布时间
        crawled_at:
          type: string
          format: date-time
          readOnly: true
          title: 抓取时间
        language:
          type: string
          title: 语言
          maxLength: 10
        tags:
          nullable: true
          title: 标签
        keywords:
          nullable: true
          title: 关键词
        sentiment_score:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          title: 情感分值
        importance_score:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          title: 重要性分值
        relevance_score:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          title: 相关性分值
        quality_score:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,2})?$
          nullable: true
          title: 质量分值
        is_processed:
          type: boolean
          title: 是否已处理
        is_duplicate:
          type: boolean
          title: 是否重复
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: 创建时间
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: 更新时间
        category:
          type: integer
          nullable: true
          title: 分类
        generated_article:
          type: integer
          nullable: true
          title: 生成的文章
      required:
      - content
      - crawled_at
      - created_at
      - id
      - published_at
      - source_url
      - title
      - updated_at
...
    TagTypeEnum:
      enum:
      - general
      - topic
      - emotion
      - category
      - location
      - person
      - event
      - technology
      - trend
      type: string
      description: |-
        * `general` - 通用
        * `topic` - 话题
        * `emotion` - 情感
        * `category` - 分类
        * `location` - 地域
        * `person` - 人物
        * `event` - 事件
        * `technology` - 技术
        * `trend` - 趋势
...
    TypeEnum:
      enum:
      - audio
      - video
      - article
      type: string
      description: |-
        * `audio` - 音频
        * `video` - 视频
        * `article` - 文章
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT